- Update `LogLevel` namespace and references across the project - Update minimum PHP version to 8.2 in `composer.json` - Add additional autoload files and QA scripts in `composer.json` - Refactor `MetricOperation` to a dedicated enum - Introduce `RequiredInitializationParams` for cleaner initialization logic - Improve log sync and metric operations with static `InternalLogger` usage
