+++ src/Metrics/BaseMetrics.php
@@ -7,6 +7,6 @@ namespace Logdash\Metrics;
 interface BaseMetrics
 {
     public function set(string $key, float $value): void;
-    
+
     public function mutate(string $key, float $value): void;
 }
