+++ src/Metrics/MetricOperation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Logdash\Metrics;
+
+enum MetricOperation: string
+{
+    case SET = 'set';
+    case CHANGE = 'change';
+}
