+++ src/Types/InitializationParams.php
@@ -10,7 +10,8 @@ class InitializationParams
         public readonly ?string $apiKey = null,
         public readonly string $host = 'https://api.logdash.io',
         public readonly bool $verbose = false
-    ) {}
+    ) {
+    }
 
     public function toRequired(): RequiredInitializationParams
     {
@@ -20,13 +21,4 @@ class InitializationParams
             verbose: $this->verbose
         );
     }
-}
-
-class RequiredInitializationParams
-{
-    public function __construct(
-        public readonly string $apiKey,
-        public readonly string $host,
-        public readonly bool $verbose
-    ) {}
 }
